+++ build_tools/circle/build_test_pypy.sh
@@ -16,9 +16,10 @@ source pypy-env/bin/activate
 python --version
 which python
 
-pip install --extra-index https://antocuni.github.io/pypy-wheels/ubuntu numpy Cython pytest
+pip install --extra-index https://antocuni.github.io/pypy-wheels/ubuntu numpy Cython pytest pytest-mpl
 pip install "scipy>=1.1.0"
-pip install scikit-learn pandas statsmodels
+pip install "scikit-learn==0.19.1"
+pip install pandas statsmodels matplotlib
 
 ccache -M 512M
 export CCACHE_COMPRESS=1
@@ -29,4 +30,4 @@ pip install -vv -e .
 
 # Pytest is known to consume lots of memory for a large number of tests,
 # and Circle 2.0 limits 4GB per container.
-python -m pytest pmdarima/ -p no:logging
+python -m pytest pmdarima/ -p no:logging --mpl --mpl-baseline-path=pytest_images