+++ pmdarima/arima/tests/test_stationarity.py
@@ -1,13 +1,15 @@
-# stationarity/seasonality tests
+# -*- coding: utf-8 -*-
+# stationarity tests
 
 from __future__ import absolute_import
 
 from pmdarima.arima.stationarity import ADFTest, PPTest, KPSSTest
-from pmdarima.arima.seasonality import CHTest
-from pmdarima.arima.utils import ndiffs, nsdiffs
+from pmdarima.arima.utils import ndiffs
+from pmdarima.utils.array import diff
 
 from sklearn.utils import check_random_state
-from numpy.testing import assert_array_almost_equal, assert_almost_equal
+from numpy.testing import assert_array_almost_equal, assert_almost_equal, \
+    assert_array_equal
 
 import numpy as np
 import pytest
@@ -15,6 +17,7 @@ import pytest
 # for testing rand of len 400 for m==365
 random_state = check_random_state(42)
 
+# TODO: redundant code with test_seasonality. Fix this in separate feature
 austres = np.array([13067.3, 13130.5, 13198.4, 13254.2, 13303.7, 13353.9,
                     13409.3, 13459.2, 13504.5, 13552.6, 13614.3, 13669.5,
                     13722.6, 13772.1, 13832.0, 13862.6, 13893.0, 13926.8,
@@ -40,31 +43,116 @@ def test_ndiffs_stationary():
     assert ndiffs(x, alpha=0.05, test='adf', max_d=2) == 0
 
 
-def test_embedding():
+@pytest.mark.parametrize("cls", (KPSSTest, PPTest, ADFTest))
+def test_embedding(cls):
     x = np.arange(5)
     expected = np.array([
         [1, 2, 3, 4],
         [0, 1, 2, 3]
     ])
 
-    assert_array_almost_equal(KPSSTest()._embed(x, 2), expected)
+    assert_array_almost_equal(cls._embed(x, 2), expected)
+
+    y = np.array([1, -1, 0, 2, -1, -2, 3])
+    assert_array_almost_equal(cls._embed(y, 1),
+                              np.array([
+                                  [1, -1, 0, 2, -1, -2, 3]
+                              ]))
+
+    assert_array_almost_equal(cls._embed(y, 2).T,
+                              np.array([
+                                  [-1, 1],
+                                  [0, -1],
+                                  [2, 0],
+                                  [-1, 2],
+                                  [-2, -1],
+                                  [3, -2]
+                              ]))
+
+    assert_array_almost_equal(cls._embed(y, 3).T,
+                              np.array([
+                                  [0, -1, 1],
+                                  [2, 0, -1],
+                                  [-1, 2, 0],
+                                  [-2, -1, 2],
+                                  [3, -2, -1]
+                              ]))
+
+    # Where K close to y dim
+    assert_array_almost_equal(cls._embed(y, 6).T,
+                              np.array([
+                                  [-2, -1, 2, 0, -1, 1],
+                                  [3, -2, -1, 2, 0, -1]
+                              ]))
+
+    # Where k == y dim
+    assert_array_almost_equal(cls._embed(y, 7).T,
+                              np.array([
+                                  [3, -2, -1, 2, 0, -1, 1]
+                              ]))
+
+    # Assert we fail when k > dim
+    with pytest.raises(ValueError):
+        cls._embed(y, 8)
 
 
-def test_kpss():
-    test = KPSSTest(alpha=0.05, null='level', lshort=True)
-    pval, is_sig = test.is_stationary(austres)
-    assert is_sig  # show it is significant
+def test_adf_ols():
+    # Test the _ols function of the ADF test
+    x = np.array([1, -1, 0, 2, -1, -2, 3])
+    k = 2
+    y = diff(x)
+    assert_array_equal(y, [-2, 1, 2, -3, -1, 5])
+
+    z = ADFTest._embed(y, k).T
+    res = ADFTest._ols(x, y, z, k)
+
+    # Assert on the params of the OLS. The comparisons are those obtained
+    # from the R function.
+    expected = np.array([1.0522, -3.1825, -0.1609, 1.4690])
+    assert np.allclose(res.params, expected, rtol=0.001)
+
+    # Now assert on the standard error
+    stat = ADFTest._ols_std_error(res)
+    assert np.allclose(stat, -100.2895)  # derived from R code
+
+
+def test_adf_p_value():
+    # Assert on the ADF test's p-value
+    p_val, do_diff = \
+        ADFTest(alpha=0.05).is_stationary(np.array([1, -1, 0, 2, -1, -2, 3]))
+
+    assert np.isclose(p_val, 0.01)
+    assert not do_diff
+
+
+@pytest.mark.parametrize('null', ('level', 'trend'))
+def test_kpss(null):
+    test = KPSSTest(alpha=0.05, null=null, lshort=True)
+    pval, do_diff = test.is_stationary(austres)
+    assert do_diff  # show it is significant
     assert_almost_equal(pval, 0.01)
 
+    # Test on the data provided in issue #67
+    x = np.array([1, -1, 0, 2, -1, -2, 3])
+    pval2, do_diff2 = test.is_stationary(x)
+
+    # We expect Trend to be significant, but NOT Level
+    if null == 'level':
+        assert not do_diff2
+        assert_almost_equal(pval2, 0.1)
+    else:
+        assert do_diff2
+        assert_almost_equal(pval2, 0.01)
+
     # test the ndiffs with the KPSS test
-    assert ndiffs(austres, test='kpss', max_d=2) == 2
+    assert ndiffs(austres, test='kpss', max_d=5, null=null) == 2
 
 
 def test_non_default_kpss():
     test = KPSSTest(alpha=0.05, null='trend', lshort=False)
-    pval, is_sig = test.is_stationary(austres)
-    assert is_sig  # show it is significant
-    assert_almost_equal(pval, 0.01)
+    pval, do_diff = test.is_stationary(austres)
+    assert do_diff  # show it is significant
+    assert np.allclose(pval, 0.01, atol=0.005)
 
     # test the ndiffs with the KPSS test
     assert ndiffs(austres, test='kpss', max_d=2) == 2
@@ -78,21 +166,50 @@ def test_kpss_corner():
 
 def test_pp():
     test = PPTest(alpha=0.05, lshort=True)
-    pval, is_sig = test.is_stationary(austres)
-    assert is_sig
-    assert_almost_equal(pval, 0.02139, decimal=5)
+    pval, do_diff = test.is_stationary(austres)
+    assert do_diff
+
+    # Result from R code: 0.9786066
+    # > pp.test(austres, lshort=TRUE)$p.value
+    assert_almost_equal(pval, 0.9786066, decimal=5)
 
     # test n diffs
-    nd = ndiffs(austres, test='pp', max_d=2)
-    assert nd == 1
+    assert ndiffs(austres, test='pp', max_d=2) == 1
+
+    # If we use lshort is FALSE, it will be different
+    test = PPTest(alpha=0.05, lshort=False)
+    pval, do_diff = test.is_stationary(austres)
+    assert do_diff
+
+    # Result from R code: 0.9514589
+    # > pp.test(austres, lshort=FALSE)$p.value
+    assert_almost_equal(pval, 0.9514589, decimal=5)
+    assert ndiffs(austres, test='pp', max_d=2, lshort=False) == 1
 
 
 def test_adf():
-    test = ADFTest(alpha=0.05, k=2)
-    pval, is_sig = test.is_stationary(austres)
-    assert not is_sig
+    # Test where k = 1
+    test = ADFTest(alpha=0.05, k=1)
+    pval, do_diff = test.is_stationary(austres)
 
-    # OLS in statsmodels seems unstable compared to the R code?...
+    # R's value: 0.8488036
+    # > adf.test(austres, k=1, alternative='stationary')$p.value
+    assert np.isclose(pval, 0.8488036)
+    assert do_diff
+
+    # Test for k = 2. R's value: 0.7060733
+    # > adf.test(austres, k=2, alternative='stationary')$p.value
+    test = ADFTest(alpha=0.05, k=2)
+    pval, do_diff = test.is_stationary(austres)
+    assert np.isclose(pval, 0.7060733)
+    assert do_diff
+
+    # Test for k is None. R's value: 0.3493465
+    # > adf.test(austres, alternative='stationary')$p.value
+    test = ADFTest(alpha=0.05, k=None)
+    pval, do_diff = test.is_stationary(austres)
+    assert np.isclose(pval, 0.3493465, rtol=0.0001)
+    assert do_diff
 
 
 def test_adf_corner():
@@ -104,57 +221,9 @@ def test_adf_corner():
     test.is_stationary(austres)
 
 
-def test_ch_test():
-    val = CHTest._sd_test(austres, 3)
-
-    # R code produces 0.07956102
-    assert_almost_equal(val, 0.07956102, decimal=7)
-    assert CHTest(m=3).estimate_seasonal_differencing_term(austres) == 0
-
-    # what if freq > 12?
-    assert_almost_equal(CHTest._sd_test(austres, 24), 4.134289, decimal=5)
-    assert CHTest(m=24).estimate_seasonal_differencing_term(austres) == 0
-    assert CHTest(m=52).estimate_seasonal_differencing_term(austres) == 0
-
-    # this won't even go thru because n < 2 * m + 5:
-    assert CHTest(m=365).estimate_seasonal_differencing_term(austres) == 0
-
-    # change the length to be longer so we can actually test the end case
-    aus_list = austres.tolist()  # type: list
-    y = np.asarray(aus_list * 10)  # type: np.ndarray
-
-    # y len is now 1760, which is > 2 * m + 5, but idk what to assert
-    CHTest(m=365).estimate_seasonal_differencing_term(y)
-
-    # what if m > 365???
-    CHTest(m=366).estimate_seasonal_differencing_term(y)
-
-
-def test_ch_base():
-    test = CHTest(m=2)
-    assert test.estimate_seasonal_differencing_term(None) == 0
-
-    # test really long m for random array
-    CHTest(m=365).estimate_seasonal_differencing_term(random_state.rand(400))
-
-
 def test_ndiffs_corner_cases():
     with pytest.raises(ValueError):
         ndiffs(austres, max_d=0)
-
-
-def test_nsdiffs_corner_cases():
-    # max_D must be a positive int
-    with pytest.raises(ValueError):
-        nsdiffs(austres, m=2, max_D=0)
-
-    # assert 0 for constant
-    assert nsdiffs([1, 1, 1, 1], m=2) == 0
-
-    # show fails for m <= 1
-    for m in (0, 1):
-        with pytest.raises(ValueError):
-            nsdiffs(austres, m=m)
 
 
 def test_base_cases():