+++ doc/whats_new.rst
@@ -7,8 +7,39 @@ What's new in pmdarima
 As new releases of pmdarima are pushed out, the following list (introduced in
 v0.8.1) will document the latest features.
 
-v1.0.0
-------
+`v1.1.0 <http://alkaline-ml.com/pmdarima/1.1.0/>`_
+--------------------------------------------------
+
+* Added ``ARIMA.plot_diagnostics`` method, as requested in `#49 <https://github.com/tgsmith61591/pmdarima/issues/49>`_
+
+* Added new arg to ``ARIMA`` constructor and ``auto_arima``: ``with_intercept`` (default is True).
+
+* New default for ``trend`` is no longer ``'c'``, it is ``None``.
+
+* Added ``to_dict`` method to ``ARIMA`` class to address `Issue #54 <https://github.com/tgsmith61591/pmdarima/issues/54>`_
+
+* ARIMA serialization no longer stores statsmodels results wrappers in the cache,
+  but bundles them into the pickle file. This solves `Issue #48 <https://github.com/tgsmith61591/pmdarima/issues/48>`_
+  and only works on statsmodels 0.9.0+ since they've fixed a bug on their end.
+
+* The ``'PMDARIMA_CACHE'`` and ``'PMDARIMA_CACHE_WARN_SIZE'`` environment variables are
+  now deprecated, since they no longer need to be used.
+
+* Added versioned documentation. All releases' doc (from 0.9.0 onward) is now available
+  at ``alkaline-ml.com/pmdarima/<version>``
+
+* Fix bug in ``ADFTest`` where ``OLS`` was computed with ``method="pinv"`` rather
+  than ``"method=qr"``. This fix means better parity with R's results. See
+  `#71 <https://github.com/tgsmith61591/pmdarima/pull/71>`_ for more context.
+
+* ``CHTest`` now solves linear regression with ``normalize=True``. This solves
+  `#74 <https://github.com/tgsmith61591/pmdarima/issues/74>`_
+
+* Python 3.7 is now supported(!!)
+
+
+`v1.0.0 <http://alkaline-ml.com/pmdarima/1.0.0/>`_
+--------------------------------------------------
 
 * **Wheels will no longer be built for Python versions < 3.5.** You may still be able to build
   from source, but support for 2.x python versions will diminish in future versions.
@@ -32,8 +63,8 @@ v1.0.0
   - ``'PYRAMID_ARIMA_CACHE_WARN_SIZE'`` will become ``'PMDARIMA_CACHE_WARN_SIZE'``
 
 
-v0.9.0
-------
+`v0.9.0 <http://alkaline-ml.com/pmdarima/0.9.0/>`_
+--------------------------------------------------
 
 * Explicitly catch case in ``auto_arima`` where a value of ``m`` that is too large may over-estimate
   ``D``, causing the time series to be differenced down to an empty array. This is now handled by